msm: camera: isp: Fix HVX streaming after crash
If user space process crashed, hvx state is not updated since it is
updated during stream off (which never gets called during stream off).
When camera is opened after crash, HVX configuration is not reset.
This leads to data routed to HVX even if its disabled. Update HVX
state during stream ON and configure HVX with enable / disable to
fix the issue.

Change-Id: I4b9cbba1261eafd83c67275c5d7dc3defbb3d990
Signed-off-by: Sreesudhan Ramakrish Ramkumar <srramku@codeaurora.org>